test: split the contract file again along the question it asks
Codex round 22. The first split relocated the bulk instead of dividing
it, and the file grew back past ~400 lines as findings landed.

The seam this time is the question each block asks. Deferrals and the
Supervisor lifecycle path answer WHERE a call belongs; expansion and
tier classification answer WHAT a run turns out to do. 324 and 114
lines, and the new file is registered in safe-core-files.json so it
actually runs.
