+++ scripts/test/safe-core-files.json
@@ -74,6 +74,7 @@
   "tests/skills/bulk-audit-contract.test.ts",
   "tests/skills/config-snapshots-contract.test.ts",
   "tests/skills/confirmation-code-terminology.test.ts",
+  "tests/skills/deferral-contract.test.ts",
   "tests/skills/energy-contract.test.ts",
   "tests/skills/grouped-change-set-contract.test.ts",
   "tests/skills/ha-cross-skill-integration.test.ts",
