+++ tests/skills/indirect-actuation-contract.test.ts
@@ -13,7 +13,6 @@ const read = (relative: string): string =>
 
 const skillDoc = read("skills/service-call/SKILL.md");
 const contextSkill = read("skills/ha-nova/SKILL.md");
-const fallbackSkill = read("skills/fallback/SKILL.md");
 const reviewSkill = read("skills/review/SKILL.md");
 const mqttSkill = read("skills/mqtt/SKILL.md");
 const adminSkill = read("skills/admin/SKILL.md");
@@ -24,80 +23,7 @@ const writeSkill = read("skills/write/SKILL.md");
 // Contract docs hard-wrap at ~72 columns, so a pinned sentence must not also
 // pin the column it happens to break at.
 const flat = (text: string): string => text.replace(/\s+/g, " ");
-
-describe("indirect actuation and owning-skill deferrals (#513)", () => {
-  describe("owning-skill deferrals cover every gated service family (#513)", () => {
-    // A generic service call must never reach a family whose owning skill
-    // carries stricter gates. Each row below exists because the owning skill
-    // enforces something this flow does not: a feature bitmask, a backup
-    // offer, an impact quantification, or a typed confirmation code.
-    it("defers every gated family to its owning skill", () => {
-      for (const [service, owner] of [
-        ["mqtt.publish", "ha-nova:mqtt"],
-        ["update.install", "ha-nova:updates"],
-        ["camera.snapshot", "ha-nova:camera"],
-        ["camera.turn_on", "ha-nova:camera"],
-        ["media_player.*", "ha-nova:media"],
-        ["notify.*", "ha-nova:notify"],
-        ["logger.set_level", "ha-nova:diagnose"],
-        ["recorder.purge", "ha-nova:maintenance"],
-        ["recorder.purge_entities", "ha-nova:maintenance"],
-        ["calendar.create_event", "ha-nova:calendar"],
-        ["todo.add_item", "ha-nova:todo"],
-        ["backup.create", "ha-nova:backup"],
-        ["conversation.process", "ha-nova:assist"],
-      ] as Array<[string, string]>) {
-        // A row may name the service outright or cover its whole domain
-        // (`camera.*`), so accept either spelling — otherwise a later PR that
-        // consolidates a family silently loses its deferral.
-        const domainWildcard = `\`${service.split(".")[0]}.*\``;
-        const row = skillDoc
-          .split("\n")
-          .find(
-            (line) =>
-              line.startsWith("|") &&
-              (line.includes(`\`${service}\``) || line.includes(domainWildcard)),
-          );
-        expect(row, `no deferral row for ${service}`).toBeTruthy();
-        expect(row, `${service} must defer to ${owner}`).toContain(owner);
-      }
-    });
-
-    it("keeps read-only response services and scene activation in this flow", () => {
-      expect(skillDoc).toContain("Read-only response services stay here");
-      expect(skillDoc).toContain("`calendar.get_events`");
-      expect(skillDoc).toContain("`todo.get_items`");
-      expect(skillDoc).toContain(
-        "`ha-nova:scene` owns scene CRUD, not activation",
-      );
-    });
-
-    it("keeps Supervisor lifecycle here rather than routing it somewhere weaker", () => {
-      // fallback tiers Apps as External, so deferring hassio.* would send the
-      // agent to a page that denies the transport works — and fallback has no
-      // disruptive tier and no self-amputation rule.
-      const rows = skillDoc
-        .split("\n")
-        .filter((line) => line.startsWith("|") && line.includes("hassio"));
-      expect(rows.length).toBeGreaterThanOrEqual(2);
-      expect(rows.join(" ")).toContain("disruptive tier");
-      // The domain is not a wildcard: restores reboot HA and updates have an
-      // owning skill, so both must route away from the generic flow.
-      expect(flat(skillDoc)).toContain("The `hassio` domain is NOT a wildcard");
-      expect(flat(skillDoc)).toContain("belong to `ha-nova:backup`, which refuses restores outright");
-      expect(flat(skillDoc)).toContain("`addon_update` belongs to `ha-nova:updates`");
-      expect(flat(skillDoc)).toContain(
-        "Refuse outright any call targeting the App that runs this Relay",
-      );
-      // An App restart takes every device that App serves offline.
-      expect(skillDoc).toContain("`hassio.addon_restart`");
-      expect(skillDoc).toContain("`hassio.addon_start`");
-      expect(flat(skillDoc)).toContain(
-        "an MQTT or Z-Wave App takes every device it serves offline while it comes back",
-      );
-    });
-  });
-
+describe("indirect actuation and tier classification (#513)", () => {
   describe("indirect actuation cannot bypass the high-consequence tier (#513)", () => {
     it("routes every indirect-run service through the shared gate", () => {
       expect(skillDoc).toContain("Indirect actuation gate");
@@ -245,7 +171,6 @@ describe("indirect actuation and owning-skill deferrals (#513)", () => {
       expect(assistSkill).toContain("including the re-run proof after an exposure fix");
     });
   });
-
   describe("sibling flows restate the access gate (#513 class sweep)", () => {
     it("excludes access-granting corrections from review Quick-Fix", () => {
       expect(reviewSkill).toContain(
@@ -291,7 +216,6 @@ describe("indirect actuation and owning-skill deferrals (#513)", () => {
       expect(contextSkill).toContain("user-account creation");
     });
   });
-
   it("reads a Template button's own action instead of trusting search/related", () => {
     const gate = flat(indirectActuation);
     // Every spelling of a script run still has to be enumerated somewhere.
@@ -309,30 +233,12 @@ describe("indirect actuation and owning-skill deferrals (#513)", () => {
     expect(gate).toContain("if you cannot read it, escalate");
     expect(gate).toContain("is an integration button whose behaviour");
   });
-
-  it("gives Supervisor lifecycle calls a path the entity flow cannot provide", () => {
-    const doc = flat(skillDoc);
-    // An App is addressed by slug; there is no entity to read before or after.
-    expect(doc).toContain("The target is an App SLUG, not an entity");
-    // Probed live: the relay passes only /api/..., and HA answers its
-    // /api/hassio/... proxy with 403 — so the slug comes from the update
-    // entity's entity_picture, and App state is simply not readable here.
-    expect(doc).toContain("The Supervisor API is NOT reachable from here");
-    expect(doc).toContain("`entity_picture` is `/api/hassio/addons/<slug>/icon`");
-    expect(doc).toContain("App state cannot be verified from here");
-    expect(doc).toContain("Never infer success from the service call returning");
-    // A host reboot takes the transport with it, so success is unobservable.
-    expect(doc).toContain("Never report success: you will not be there to see it");
-    expect(doc).toContain("needs physical access to come back");
-  });
-
   it("does not let a targetless reload skip the scan for lack of a target", () => {
     const gate = flat(indirectActuation);
     expect(gate).toContain("A targetless reload of a trigger-source domain");
     expect(gate).toContain("Its effect set is every entity of that domain");
     expect(gate).toContain('"No target" is not evidence of no impact');
   });
-
   it("does not read an empty consumer scan as an absence of consumers", () => {
     const gate = flat(indirectActuation);
     expect(gate).toContain('A zero-hit scan is NOT "no consumers"');
@@ -348,7 +254,6 @@ describe("indirect actuation and owning-skill deferrals (#513)", () => {
     expect(gate).toContain("the unread thing IS the consumer");
     expect(gate).toContain("an installed-but-unreadable Node-RED escalates");
   });
-
   it("does not render the ordinary confirmation menu on the typed tier", () => {
     const doc = flat(skillDoc);
     // A gate that assigns a tier and then offers "apply" has assigned nothing.
@@ -356,7 +261,6 @@ describe("indirect actuation and owning-skill deferrals (#513)", () => {
     expect(doc).toContain("the only accepted answer is the exact `confirm:<token>`");
     expect(doc).toContain("including when the tier came from an EXPANDED member");
   });
-
   it("lets the read state, not the alias, decide whether a call is a run", () => {
     const doc = flat(skillDoc);
     expect(doc).toContain("Entering the gate is not the same as being a run");
@@ -365,7 +269,6 @@ describe("indirect actuation and owning-skill deferrals (#513)", () => {
     // The Flow entry must not pre-classify the alias as a run.
     expect(doc).not.toContain("on `script.open_door` is a script run");
   });
-
   it("fails closed on an unresolved cover target", () => {
     const gate = flat(indirectActuation);
     // A blind and a garage door take the same call; only device_class
@@ -373,14 +276,12 @@ describe("indirect actuation and owning-skill deferrals (#513)", () => {
     expect(gate).toContain("a blind and a garage door take the same call");
     expect(gate).toContain("an unresolved cover target fails CLOSED");
   });
-
   it("keeps native automation lifecycle services out of the run class", () => {
     const gate = flat(indirectActuation);
     // automation.turn_on enables; only automation.trigger executes.
     expect(gate).toContain("`automation.turn_on|turn_off|toggle`");
     expect(gate).toContain("only flip whether it may fire later");
   });
-
   it("treats a cycle between fully-read members as resolved", () => {
     const gate = flat(indirectActuation);
     // Escalating every cut cycle puts harmless retry loops on the typed tier.
@@ -388,17 +289,11 @@ describe("indirect actuation and owning-skill deferrals (#513)", () => {
     expect(gate).toContain("do not become access-capable by looping");
     expect(gate).toContain("A cycle through a member you could NOT read is still unresolved");
   });
-
   it("applies the Template-carrier rule to switches as well as buttons", () => {
     const gate = flat(indirectActuation);
     expect(gate).toContain("A `button` or `switch` is the exception");
     expect(gate).toContain("A Template switch defines its own");
   });
-
-  it("defers clearing completed to-do items to the owning skill", () => {
-    expect(flat(skillDoc)).toContain("`todo.remove_completed_items`");
-  });
-
   it("classifies the device-action form, not only the service form", () => {
     const gate = flat(indirectActuation);
     // The automation editor emits domain/type/device_id for a device-picked
@@ -410,19 +305,16 @@ describe("indirect actuation and owning-skill deferrals (#513)", () => {
     expect(gate).toContain("reads back as `use_blueprint` with inputs, not");
     expect(gate).toContain("It is an unresolved branch, not an empty one");
   });
-
   it("fails closed for every cover action that can open", () => {
     const gate = flat(indirectActuation);
     expect(gate).toContain("`open_cover`, `toggle` (a closed garage opens), and `set_cover_position`");
     expect(gate).toContain("Only `close_cover` is safe on an unknown target");
   });
-
   it("carries an owning skill's typed tier through an expansion", () => {
     const gate = flat(indirectActuation);
     expect(gate).toContain("carries that tier into the run");
     expect(gate).toContain("Physical access is not the only reason a member is gated");
   });
-
   it("does not read a local-presence check as proof of no consumer", () => {
     const gate = flat(indirectActuation);
     expect(gate).toContain("Presence detection sees LOCAL installs only");
